Add weighted_add method to BattingSummary
The type conversions here are ugly. I can either not allow floats for
the weights or make the summary use f32/f64.

I don't want to restrict the weights to integers because it's limiting.
I could support floats but internally convert to integers. More likely
there will be a new summary type that saves the data as f64.