Split up projection code
Ported the projection code to be in two modules: databank and
projection. The calculating of the projection now happens in the
projection module with its own struct. The manipulation of season data
happens in the databank module.

The original code is still there. Once the projections can be written
out from the new projection module, then the original code will be
completely removed.