Support a player not appearing in a season
Expand the PlayerType by adding a fourth option: NoAppearances. This
means the player did not appear at all that season.

There is one caveat for pitchers: they can have 0 IP but still have >0
batters faced. Including the number of batters faced remains to be done.
It might not be that important because if that player did not have any
PAs in the season, they will still count as a pitcher.